Update: Added modifiers and events
I replaced the `require` statements in the `removeBillboard` and `refurbish_Billboardurl` with a `modifier` checkOwner which will check if the person calling this function is the owner of billboard or not. This reduces the code in the contract. Also added some events which can be rendered in the frontend using useEffect hooks or document.ready in Javascript.